# Stack-based reducer