--- a/packages/components/src/components/SquiggleChart.tsx
+++ b/packages/components/src/components/SquiggleChart.tsx
@@ -1,14 +1,21 @@
 import * as React from "react";
-import { useSquiggle, SquiggleArgs } from "../lib/hooks/useSquiggle.js";
-import { SquiggleViewer, SquiggleViewerProps } from "./SquiggleViewer/index.js";
+
+import { RefreshIcon } from "@quri/ui";
+
+import { SquiggleArgs, useSquiggle } from "../lib/hooks/useSquiggle.js";
 import { getValueToRender } from "../lib/utility.js";
+import { SquiggleViewer, SquiggleViewerProps } from "./SquiggleViewer/index.js";
 
 type Props = SquiggleArgs & Omit<SquiggleViewerProps, "result">;
 
 export const SquiggleChart: React.FC<Props> = React.memo((props) => {
   // TODO - split props into useSquiggle args and SquiggleViewer args would be cleaner
   // (but `useSquiggle` props are union-typed and are hard to extract for that reason)
-  const resultAndBindings = useSquiggle(props);
+  const [resultAndBindings] = useSquiggle(props);
+
+  if (!resultAndBindings) {
+    return <RefreshIcon className="animate-spin" />;
+  }
 
   const valueToRender = getValueToRender(resultAndBindings);