--- a/packages/components/src/components/SquiggleErrorAlert.tsx
+++ b/packages/components/src/components/SquiggleErrorAlert.tsx
@@ -1,43 +1,77 @@
-import { SqError, SqFrame } from "@quri/squiggle-lang";
-import React from "react";
+import { FC, PropsWithChildren } from "react";
+
+import {
+  SqCompileError,
+  SqError,
+  SqFrame,
+  SqRuntimeError,
+} from "@quri/squiggle-lang";
+
 import { ErrorAlert } from "./Alert.js";
+import { useViewerContext } from "./SquiggleViewer/ViewerProvider.js";
 
 type Props = {
   error: SqError;
 };
 
-const StackTraceFrame: React.FC<{ frame: SqFrame }> = ({ frame }) => {
+const LocationLine: FC<{
+  location: NonNullable<ReturnType<SqFrame["location"]>>;
+}> = ({ location }) => {
+  const { editor } = useViewerContext();
+
+  const findInEditor = () => {
+    editor?.scrollTo(location.start.offset);
+  };
+
+  return (
+    <span className="cursor-pointer hover:text-red-900" onClick={findInEditor}>
+      at line {location.start.line}, column {location.start.column}
+    </span>
+  );
+};
+
+const WithHeader: FC<PropsWithChildren<{ header: string }>> = ({
+  header,
+  children,
+}) => (
+  <div>
+    <div className="font-medium">{header}</div>
+    <div className="ml-4">{children}</div>
+  </div>
+);
+
+const StackTraceFrame: FC<{ frame: SqFrame }> = ({ frame }) => {
   const location = frame.location();
   return (
     <div>
-      {frame.name()}
-      {location
-        ? ` at line ${location.start.line}, column ${location.start.column}`
-        : ""}
+      {frame.name()} {location ? <LocationLine location={location} /> : ""}
     </div>
   );
 };
 
-const StackTrace: React.FC<Props> = ({ error }) => {
+const StackTrace: FC<{ error: SqRuntimeError }> = ({ error }) => {
   const frames = error.getFrameArray();
   return frames.length ? (
-    <div>
-      <div className="font-medium">Stack trace:</div>
-      <div className="ml-4">
-        {frames.map((frame, i) => (
-          <StackTraceFrame frame={frame} key={i} />
-        ))}
-      </div>
-    </div>
+    <WithHeader header="Stack trace:">
+      {frames.map((frame, i) => (
+        <StackTraceFrame frame={frame} key={i} />
+      ))}
+    </WithHeader>
   ) : null;
 };
 
-export const SquiggleErrorAlert: React.FC<Props> = ({ error }) => {
+export const SquiggleErrorAlert: FC<Props> = ({ error }) => {
   return (
     <ErrorAlert heading="Error">
       <div className="space-y-4">
         <div className="whitespace-pre-wrap">{error.toString()}</div>
-        <StackTrace error={error} />
+        {error instanceof SqRuntimeError ? (
+          <StackTrace error={error} />
+        ) : error instanceof SqCompileError ? (
+          <WithHeader header="Location:">
+            <LocationLine location={error.location()} />
+          </WithHeader>
+        ) : null}
       </div>
     </ErrorAlert>
   );