--- a/packages/components/src/components/SquigglePlayground/RunControls/RunMenuItem.tsx
+++ b/packages/components/src/components/SquigglePlayground/RunControls/RunMenuItem.tsx
@@ -1,18 +1,16 @@
-import { PlayIcon } from "@quri/ui";
 import React from "react";
-
 import { RefreshIcon } from "@heroicons/react/solid/esm/index.js";
+
+import { PlayIcon } from "@quri/ui";
 
 import { isMac } from "../../../lib/utility.js";
 import { MenuItem } from "../MenuItem.js";
 import { RunnerState } from "./useRunnerState.js";
 
-export const RunMenuItem: React.FC<RunnerState> = ({
+export const RunMenuItem: React.FC<RunnerState & { isRunning: boolean }> = ({
   run,
-  code,
-  renderedCode,
-  isRunning,
   autorunMode,
+  isRunning,
 }) => {
   const showAsRunning = !autorunMode && isRunning;
   const text = `Run (${isMac() ? "Cmd+Enter" : "Ctrl+Enter"})`;