--- a/packages/components/src/components/SquigglePlayground/RunControls/useRunnerState.ts
+++ b/packages/components/src/components/SquigglePlayground/RunControls/useRunnerState.ts
@@ -1,22 +1,15 @@
-import { useLayoutEffect, useReducer, useRef } from "react";
+import { useLayoutEffect, useReducer } from "react";
 
 type InternalState = {
   autorunMode: boolean;
   renderedCode: string;
-  // "prepared" is for rendering a spinner; "run" for executing squiggle code; then it gets back to "none" on the next render
-  runningState: "none" | "prepared" | "run";
   executionId: number;
-  startTime?: number;
-  totalTime?: number;
 };
 
 const buildInitialState = (code: string): InternalState => ({
   autorunMode: true,
   renderedCode: "",
-  runningState: "none",
   executionId: 0,
-  startTime: undefined,
-  totalTime: undefined,
 });
 
 type Action =
@@ -26,14 +19,8 @@ type Action =
       code: string;
     }
   | {
-      type: "PREPARE_RUN";
-    }
-  | {
       type: "RUN";
       code: string;
-    }
-  | {
-      type: "STOP_RUN";
     };
 
 const reducer = (state: InternalState, action: Action): InternalState => {
@@ -43,76 +30,35 @@ const reducer = (state: InternalState, action: Action): InternalState => {
         ...state,
         autorunMode: action.value,
       };
-    case "PREPARE_RUN":
-      return {
-        ...state,
-        runningState: "prepared",
-      };
     case "RUN":
       return {
         ...state,
-        runningState: "run",
         renderedCode: action.code,
         executionId: state.executionId + 1,
-        startTime: Date.now(),
-      };
-    case "STOP_RUN":
-      return {
-        ...state,
-        runningState: "none",
-        totalTime: state.startTime && Date.now() - state.startTime,
       };
   }
 };
 
 // For convenience, this type:
 // 1. Contains all necessary data for the playground;
-// 2. Matches the Props shape of RunControls comopnent.
+// 2. Matches the Props shape of RunControls component.
 export type RunnerState = {
   run: () => void;
   autorunMode: boolean;
   code: string;
   renderedCode: string;
-  isRunning: boolean;
   executionId: number;
-  executionTime: number | undefined;
   setAutorunMode: (newValue: boolean) => void;
 };
 
 export function useRunnerState(code: string): RunnerState {
   const [state, dispatch] = useReducer(reducer, buildInitialState(code));
-  const timeoutSetRef = useRef(false); // Ref to track if timeout is already set.
-
-  useLayoutEffect(() => {
-    const onFirstExecution = () => state.executionId === 0;
-    if (state.runningState === "prepared" && !timeoutSetRef.current) {
-      timeoutSetRef.current = true;
-      // this is necessary for async playground loading - otherwise it executes the code synchronously on the initial load
-      // (it's surprising that this is necessary, but empirically it _is_ necessary, both with `useEffect` and `useLayoutEffect`)
-      setTimeout(
-        () => {
-          dispatch({ type: "RUN", code });
-          timeoutSetRef.current = false; // Reset after dispatch.
-        },
-        // We want to delay this until after the editor renders, so that the editor shows first.
-        // 50md is often enough to do this. Later, when this is run in a web worker, we can remove.
-        onFirstExecution() ? 50 : 0
-      );
-    } else if (state.runningState === "run") {
-      dispatch({ type: "STOP_RUN" });
-    }
-  }, [state.runningState, code]);
 
   const run = () => {
-    // The rest will be handled by dispatches above on following renders, but we need to update the spinner first.
-    dispatch({ type: "PREPARE_RUN" });
+    dispatch({ type: "RUN", code });
   };
 
-  if (
-    state.autorunMode &&
-    state.renderedCode !== code &&
-    state.runningState === "none"
-  ) {
+  if (state.autorunMode && state.renderedCode !== code) {
     run();
   }
 
@@ -121,9 +67,7 @@ export function useRunnerState(code: string): RunnerState {
     autorunMode: state.autorunMode,
     code,
     renderedCode: state.renderedCode,
-    isRunning: state.runningState !== "none",
     executionId: state.executionId,
-    executionTime: state.totalTime,
     setAutorunMode: (newValue: boolean) => {
       dispatch({ type: "SET_AUTORUN_MODE", value: newValue, code });
     },