… .changeset/little-buckets-watch.md
… .changeset/nasty-pumas-hear.md
… packages/components/src/components/SquiggleViewer/ExpressionViewer.tsx
… packages/components/src/components/SquiggleViewer/VariableBox.tsx
… packages/components/src/components/SquiggleViewer/ViewerProvider.tsx
… packages/components/src/components/SquiggleViewer/index.tsx
… packages/components/src/components/SquiggleViewer/utils.ts
… packages/components/src/lib/hooks/useSquiggle.ts
… packages/components/src/lib/utility.ts
… packages/components/test/basic.test.tsx
… packages/components/test/cleanup.test.tsx
… packages/components/test/multiplot.test.tsx
… packages/components/test/viewProps.test.tsx
… packages/hub/src/relative-values/values/ModelEvaluator.ts
… packages/prettier-plugin/src/shared.ts
… packages/relative-values/src/values/ModelEvaluator.ts
… packages/squiggle-lang/__tests__/Jstat_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_imports_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_1_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_2_multisource_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_3_imports_test.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_4_injecting_user_values.ts
… packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_5_calling_functions.ts
… packages/squiggle-lang/__tests__/ast/parse_test.ts
… packages/squiggle-lang/__tests__/ast/toExpression_test.ts
… packages/squiggle-lang/__tests__/helpers/reducerHelpers.ts
… packages/squiggle-lang/__tests__/library/dist_test.ts
… packages/squiggle-lang/__tests__/library/examples_test.ts
… packages/squiggle-lang/__tests__/library/mixture_test.ts