This PR makes the interpreter significantly faster, by resolving symbols during ast->expression stage.

TODO:
- fix tests
- stack traces for undefined symbols 
- remove `value.ast` support from `components/`

Benchmark results after stage one (before: develop branch, after: stack-based reducer without value.ast branch):
bench-map: 2.6s -> 1.6s
bench-map-reduce: 9.9s -> 3.4s
GUCEM: 41s -> 31s