@berekuk 2023-06-24
This PR also includes:
- "find in editor", reimplemented with static analysis (supersedes #1921)
- "find in viewer" and other code from #1922 
- also, stacktrace lines are now clickable and open the line in editor