@berekuk 2023-06-24
Important note: since we resolve symbols during compile-time now, old code that mentioned unknown symbols in unused functions won't run anymore.

Example: [before](https://preview.squiggle-language.com/playground#code=eNqrVkpJTUsszSlxzk9JVbJSStPQVLBVqIjJi8kzVKoFAJsdCYw%3D), [after](https://squiggle-website-git-reducer-impr-49ec0b-quantified-uncertainty.vercel.app/playground#code=eNqrVkpJTUsszSlxzk9JVbJSStPQVLBVqIjJi8kzVKoFAJsdCYw%3D).

This is how it was in the old lisp interpreter, btw. But it might break the code that was written after 0.4.0; a few GUCEM models would need to be patched.

For example, [toy-world](https://squigglehub.org/users/berekuk/models/toy-world/revisions/TW9kZWxSZXZpc2lvbjpjbGhzN3ZxZnUwMDE4bDQwODMwZ2FwY2Rw) used `factorial` instead of `Danger.factorial`, and `toSamplesSet` instead of `SampleSet.fromDist`, in utils functions. (I fixed this in the latest toy-world revision on squigglehub.)