@OAGr 2023-06-25
The speed performance gain seems solid, happy to see that.